Create icarProductionPurposeType.json
Define the production purpose for which animals are bring bred, kept, or traded.
Note that breeding and selling elite animals, or trading live animals are not considered - rather the end purpose such as meat, milk, wool. Defined in terms of UNSPC (Service and Product Code) but using short names. Alternative would be to use the UNSPC code directly, but this is not as human-readable.