Change validation of birth period.
InventoryClassification birth period previously specified the birth period as an (RFC3339) icarDateType. But in fact, per the description it is an ISO8601 interval.
Changed the type to string, and added a pattern which matches a simple ISO8601 date interval (no time component) and the year and month options as per the description.