Revert "allOf" around referenced schemas
Reverting the "allOf" we added around "$ref" schema references in exampleUrlScheme.json and feedURLScheme.json.

These were added in PR #558 and removed linter warnings in the IDE - but broke the bundling script.

Because we want to release ADE 1.5 we will revert these and address the bunding or the linting warning later.